fipsprov: make X25519 and X448 internal only
As NIST does not like X25519/X448, we do not want to export it publicly for use. However, we still want those to be enabled for use in hybrid such as MLKEM768X25519. Make internal fips queries return the whole list, but for public fips queries skip over the two and do not surface them as available.

Signed-off-by: Łukasz 'sil2100' Zemczak <lukasz.zemczak@chainguard.dev>